+++ monorepo-builder.php
@@ -3,9 +3,8 @@
 declare(strict_types=1);
 
 use Symplify\MonorepoBuilder\Config\MBConfig;
-use Symplify\MonorepoBuilder\Release\ReleaseWorker\PushTagReleaseWorker;
+use Symplify\MonorepoBuilder\Release\ReleaseWorker\AddTagToChangelogReleaseWorker;
 use Symplify\MonorepoBuilder\Release\ReleaseWorker\SetCurrentMutualDependenciesReleaseWorker;
-use Symplify\MonorepoBuilder\Release\ReleaseWorker\TagVersionReleaseWorker;
 use Symplify\MonorepoBuilder\Release\ReleaseWorker\UpdateReplaceReleaseWorker;
 
 return static function (MBConfig $config): void {
@@ -13,13 +12,11 @@ return static function (MBConfig $config): void {
     $config->disableDefaultWorkers();
 
     $config->workers([
+        // change log
+        AddTagToChangelogReleaseWorker::class,
         // update root composer json
         UpdateReplaceReleaseWorker::class,
         // update child composer json
         SetCurrentMutualDependenciesReleaseWorker::class,
-        // local tag
-        TagVersionReleaseWorker::class,
-        // push tag
-        PushTagReleaseWorker::class,
     ]);
 };